refactor: extract framework mappings
